+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.1
+
+- Make `network.test` return ms if success
+- Make `network.isIP` return the ip type if success
+
 ## 1.1.0
 
 - Add `network`